requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
celery==5.3.4
flask-redis==0.4.0
redis==5.0.1
python-socketio[redis]==5.10.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
ddgs==3.9.6
googlesearch-python==1.2.3
python-dotenv==1.0.0
eventlet==0.33.3
gevent==23.9.1
aiohttp==3.9.1
python-socketio==5.10.0
flower==2.0.1